changing main.py to app.py
